(stable) When parsing HTML in remarkup, always start a new tag when encountering "<"
Summary:
Depends on D20630. See <https://discourse.phabricator-community.org/t/bug-in-src-parser-html-phutildomnode-php/2882/>. Currently, we parse `<x>1<2</x>` as an `<x>` tag containing a `<2</x>` tag.

A more expected interpretation of this is a single `<x>` tag containing the content `1<2`.

When we encounter a `<`, always start a new tag, even if we're currently in a tag.

Test Plan: Added a test and made it pass.

Reviewers: amckinley

Reviewed By: amckinley

Differential Revision: https://secure.phabricator.com/D20631